TOSCA generator translation for individual TOSCA type.
Each TOSCA abstract type is handled by a separate generator. These contain the
necessary logic to implement the abstract type as concrete Heat resource type.
In this set, only the server type requires custom handling.
All the other types are software config and are handled by the super class.
The WordPress type is a custom type and will be moved to a separate place

Partially implements blueprint heat-translator-tosca
Change-Id: Ida687e6eed53e1edfdc611bc3d27704d9d9f3d67
O