Main translation code to handle parameters
Changes to the backend Heat generator to process the parameters
from the TOSCA input section, which would be passed in from the
CLI or accept the default value.  Thi also handles the similar
get_attribute function in the output section.
Add handler for the nodejs type, needed for the monitoring use
case.  Fix the test_blockstorage unit test to match the new
expected generator output, due to a change in the Heat Nova
server resource type.

Change-Id: Id73aba9ebaa6828c676a7a2afdee7ec6ef2579d3