New custom type, unit test and template for use case
Add TOSCA template for testing:  this represents the first part
of the monitoring use case, deploying mongodb and nodejs onto
2 servers.  A nodejs application is also installed and started.
For testing, the nodejs application is set as sample, but for real
deployment, change it to some real application like:
https://github.com/paypal/rest-api-sample-app-nodejs.git
The new unit test will translate to Heat and validate the output.
Also include change for the keyword input -> inputs in the
template.

Change-Id: Ia2eabe030669fc0b0f52a9c2c1cc754a4b68a394