Add a debug env for tox
Running a test with the pdb debugger was difficult because
tox captures output and causes the pdb prompt to quit.

Tips for how to run with debug are provided here:
https://wiki.openstack.org/wiki/Testr#Debugging_.28pdb.29_Tests

This change puts these commands into a debug env in tox.ini so
you can do a command like

tox -e debug

and when it hits your breakpoint you'll get the debug prompt.

Change-Id: I84b2df606db17209706d12f6f6a03e7812210d53