Update set of scripts for mongodb and nodejs
This set of scripts is used in the first 2 nodes for the monitoring
use case.  They install mongodb and nodejs, configure and start them.
The config script for nodejs is designed for the paypal nodejs app:
https://github.com/paypal/rest-api-sample-app-nodejs.git
It configures the link to the mongodb database for the app.
The scripts have been validated by deploying to 2 VM's using Heat
templates.

Implements: blueprint translate-tosca-monitoring-usecase
Change-Id: I08d20ed40bbe23a2193ba1892d6d23b4bd7804f6