Add test templates for topology template
The templates is used for testing topology template functions. They
contain most of the features provided by topology template. But still
need improvement as the development of the TOSCA spec going on.
subsystem.yaml is used to test elements in topology template.
system.yaml is used to test substitution mapping. For it is not well
defined currently, it will be implemented later.
definition.yaml provides definitions required by the above two files.

partially implements: bp tosca-topology-template

Change-Id: I73006bb2880a69d20b647db750d806c87d0edfeb