Add handler for ConnectsTo relationship
The ConnectsTo relationship in a node can have an interface that
contains the Configure relationship.  This interface would have
a lifecycle operation associated with it, something like a
configuration script requiring some input parameters.
In the Heat template, each of these lifecycle operation would
be implemented as a software config node.
This patch adds the handler in the main translation loop to
process the ConnectsTo relationship.

Partially Implements: blueprint translate-tosca-monitoring-usecase
Change-Id: I5ca4b935cc8189ac46dca0d24a9aa8b40133928c