Handle URLs for input templates and imports
Allow input templates and imported custom types to be provided as URLs
with TOSCA-Parser auto-detecting the type (file vs URL), and add
necessary unit tests.

Note, since for some test cases currently the required file does not
exist on github, we temporarily use files hosted somewhere else. Once
the patch is merged a follow-on patch will be submitted to fix that
issue and use URLs of the new files that are submitted with this patch.

Partial-Bug: #1340748
Partially Implements: blueprint tosca-namespaces

Change-Id: Idec0318fa456ebccd552f67726eee8905714aa91