Fix backward compatibility issue
In the earlier version of TOSCA Parser the call to entry class TOSCATemplate has two
arguments, file path and parameters where parameters arg is optional. A new arg has
been added in parser as a flag to check if the path is file or not. The old way of
calling parser (e.g. ToscaTemplate(path, '')) will not work with it. Should change the
constructor signature per patch.

Change-Id: Ia543bcaa6967ae68863266faef9689b1e27f0bf7