Full validation of TOSCA templates
Modify the exception handling in TOSCA Parser so that all possible
validation errors are captured during parsing (instead of stopping
the parsing when the first error is encountered) and reported when
parsing is done.

Reporting the line number of each error will be implemented in a
separate patch.

Change-Id: I032c371fe0a73cdf901599ed4ae113f9157d8a1b
Partially-Implements: blueprint tosca-tpl-cli-validator