Add expected path for built in template
The template chnaged herewith is built in and fails parsing if provided
imports path is not an absolute path, which is expected but can be confusing
to user. Add template description to aid with the requirement and also modify
the path to a path close to possible valid path into user environment, which
can be created cloning the project into tmp dir.

Change-Id: Ib47cb975009ec873e8fdeed00fbc82c20baefcb4