Remove duplicate test
There is already a test to check whether URL is valid one. This particular
test which is removed here fails tox when run with fresh git clone on
systems where internet service provider redirects url with incorrect syntax.
Such a redirect can be fixed by user by modifying the network setting but
it can not be expected always. In order not to create any confusion and avoid
unpleasant experience running tox, it's worth while to remove this test.

Change-Id: I8b12c5f9d7cfd1787b8bc1a2ddb05350995f537e