Implement substitution mappings
Add validation for substitute_mappings and a test example.
The substitution is responsibility of an orchestrator so
work needs to be done in heat-translator to translate
substitute_mapping in HOT.

Partially example of substitute_mappings is given in
the test/data/topology_template/subsystem.yaml

Refer to the spec section 'Using node template substitution
for model composition' and related sections to read detail
on the substitute_mappings.

The related heat-translator blueprint is
https://blueprints.launchpad.net/heat-translator/+spec/
nested-templates

Partially-Implements: blueprint tosca-substitution-mappings

Co-Authored-By: Sahdev Zala <spzala@us.ibm.com>

Change-Id: I7baadef4f0b4c7cd426f710f4c74c198d02bc030
Signed-off-by: shangxdy <shang.xiaodong@zte.com.cn>