Support "Implementation" definition as Artifact Name
In the Node templates interfaces definition, Tosca
parser only recognizes implementation as a file and
not as an artifact name. ETSI SOL definition provides
that Implementation may be an artifact name, and that
the artifact definition will specify the artifact
type and the file. This modification will allow both
artifact name and file to be placed as value in
"Implementation" definition.

This is essential to allow onboarding of VNFD
packages containing the Management Driver definition
that specifies both Implementation definition value
as the Artifact Name.

https://review.opendev.org/#/c/740896/

Change-Id: I192d36499e7b731bbefaf3845f23f85d1e649ca5