import "collections.abc" explicitly
The Collections Abstract Base Classes is moved to the
collections.abc module, which was earlier part of the
collections module[1].

The python3 reports a warning for the same-
DeprecationWarning: Using or importing the ABCs from
'collections' instead of from 'collections.abc' is
deprecated since Python 3.3, and in 3.10 it will stop working.

This patch updates to import collection.abc module explicitly.

[1] https://docs.python.org/3/library/collections.abc.html

Change-Id: I18ba70140f9b2b53f2e97e70f64973af9d8953be