Support importing def from local instead of url
Sometimes importing a nested definition from remote server is failed
for the server is busy. Although we usually can download the definition
before and change the path as local then, cannot for a case if the
contents of the template is fixed and it should not be changed.

This update is avoid such a case as a remedy. The url can be replaced
with any other path by using arg `local_defs` like as below. In this
case, the definition is imported from local path instead of url.

    local_defs = {"https://example.org/orig": "/path/to/local/def"}
    tosca = ToscaTemplate(template, local_defs=local_defs}

Closes-Bug: #2006796
Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>
Change-Id: Iac88def836970cd976396452ffb9c3d276263d3e