Support for updating tosca-parser with TOSCA1.3
tosca-parser has had some updates to bring it into partial
compliance with the TOSCA Simple Profile for YAML version 1.3.

The implementation for the following updated items has been added:

The keyname of the Trigger definition was changed
'key_schema' was added to the keynames of the Property definition
'key_schema' was added to the keynames of the Parameter definition
The grammar for 'operation_definitions' in Interface Type changed.
'notifications' was added to the keynames of the Interface definition

Change-Id: I59f2c4b3f565a290e0f22d3648c9a2618dc1bc4e
Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>
Co-Authored-By: Kaori Mitani <mitani.kaori@fujitsu.com>